requirements.txt
streamlit
playwright
requests
python-dotenv
